Mostly worked on entities in the game. Finally exams are over!

+++ COMP20050-2008/groups/We_Are_Scientists/Thrust/src/thrust/entities/about/Fuelable.java
@@ -39,7 +39,7 @@ public interface Fuelable {
    */
   //@ requires 0 <= the_fuel_content & the_fuel_content <= maximum_fuel();
   //@ ensures fuel() == the_fuel_content;
-  void set_fuel_content(float the_fuel_content);
+  void set_fuel_content(int the_fuel_content);
 
   /**
    * @param the_fuel_change Change your fuel content by this many units.
@@ -50,7 +50,7 @@ public interface Fuelable {
     @             (fuel() == maximum_fuel()) :
     @           fuel() == \old(fuel() + the_fuel_change)));
     @*/
-  void change_fuel_content(float the_fuel_change);
+  void change_fuel_content(int the_fuel_change);
 
   //@ invariant (* Fuel content is always non-negative and finite. *);
   //@ invariant 0 <= fuel();
